requirements.txt
alembic==1.8.1
async-timeout==4.0.2
black==22.10.0
captcha==0.4
cfgv==3.3.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
identify==2.5.7
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
redis==4.3.4
SQLAlchemy==1.4.42
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
virtualenv==20.16.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
